feat(system): 将 system_settings.owner_id 字段类型从 Integer 改为 BigInteger

+++ core/models/system.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from sqlalchemy import String, Integer
+from sqlalchemy import String, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base
@@ -12,8 +12,9 @@ class SystemSetting(Base):
 
     owner_id=0 为全局设置，>0 为按账号隔离的设置。
     """
+
     __tablename__ = "system_settings"
 
     key: Mapped[str] = mapped_column(String(100), primary_key=True)
-    owner_id: Mapped[int] = mapped_column(Integer, primary_key=True, default=0)
+    owner_id: Mapped[int] = mapped_column(BigInteger, primary_key=True, default=0)
     value: Mapped[Optional[str]] = mapped_column(String(500))
